Apply renaming, fix css issues
