+++ compass/src/components/Application/ApplicationDetails/ApplicationDetailsScenarios/ApplicationScenarioModal.container.js
@@ -1,7 +1,7 @@
 import { compose, graphql } from 'react-apollo';
 import AssignScenarioModal from './../../../Shared/AssignScenarioModal/AssignScenarioModal.component';
-import { GET_SCENARIOS } from '../../../Shared/gql';
-import { SET_APPLICATION_SCENARIOS } from '../../gql';
+import { GET_SCENARIOS } from './../../../Shared/gql';
+import { SET_APPLICATION_SCENARIOS } from './../../gql';
 
 export default compose(
   graphql(GET_SCENARIOS, {
@@ -9,7 +9,7 @@ export default compose(
   }),
   graphql(SET_APPLICATION_SCENARIOS, {
     props: props => ({
-      setScenarios: async (applicationId, scenarios) => {
+      updateScenarios: async (applicationId, scenarios) => {
         await props.mutate({
           variables: {
             id: applicationId,
