+++ compass/src/components/Shared/AssignScenarioModal/style.scss
@@ -14,7 +14,6 @@
 }
 
 .fd-popover {
-  margin: 0 0 20px 0;
   width: 100%;
 
   :last-child {
@@ -38,6 +37,5 @@
     width: 100%;
     justify-content: space-between;
     justify-items: center;
-    padding-top: 3px;
   }
 }
